+++ server/src/Recorder/Elements/BattleUpdate.cs
@@ -2,7 +2,7 @@ using System.Text.Json;
 using System.Text.Json.Nodes;
 using System.Text.Json.Serialization;
 
-namespace Thuai.GameServer.Recorder;
+namespace Thuai.Server.Recorder;
 
 public record BattleUpdate : IRecord
 {
